% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBlockDataframe.R
\name{CreateBlockDataframe}
\alias{CreateBlockDataframe}
\title{Extract block data from json file}
\usage{
CreateBlockDataframe(d)
}
\arguments{
\item{json}{A json file imported via jsonlite::read_json()}
}
\value{
A dataframe with task data extracted from json
}
\description{
Internal function for extracting block data from json files exported from
MRL1/MRL2/VMR/MRV tasks.
}
