% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSubjectDataframe.R
\name{CreateSubjectDataframe}
\alias{CreateSubjectDataframe}
\title{Extract subject data from json file}
\usage{
CreateSubjectDataframe(json)
}
\arguments{
\item{json}{A json file imported via jsonlite::read_json()}
}
\value{
A dataframe with subject data extracted from json
}
\description{
Internal function for extracting block data from json files exported from
MRL1/MRL2/VMR/MRV tasks.
}
