% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractUserData.R
\name{ExtractUserData}
\alias{ExtractUserData}
\title{Extract user data from json file}
\usage{
ExtractUserData(json)
}
\arguments{
\item{json}{A json file imported via jsonlite::read_json()}
}
\value{
A dataframe with subject information extracted from json
}
\description{
Internal function for extracting user data from json files exported from
MRL1/MRL2/VMR/MRV tasks.
}
